backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
motor==3.3.2
pymongo==4.6.1
onnxruntime==1.16.3
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-generativeai==0.3.2
